requirements.txt
transformers
sentence_transformers
pypdf
langchain-together
openai
pinecone-client
langchain
langchain_community
langchain-openai
langchain-pinecone
tiktoken
typing_extensions
pandas
numpy
